Move const int value code to utils
